# docs/swagger.yaml
basePath: /v1
definitions:
  contract.Metadata:
    properties:
      current_page:
        type: integer
      first_page:
        type: integer
      last_page:
        type: integer
      page_size:
        type: integer
      total_records:
        type: integer
    type: object
  contract.ReqCartCreate:
    properties:
      game_id:
        minimum: 1
        type: integer
      quantity:
        minimum: 1
        type: integer
    required:
    - game_id
    - quantity
    type: object
  contract.ReqCartUpdate:
    properties:
      quantity:
        minimum: 1
        type: integer
    required:
    - quantity
    type: object
...
  contract.ResCartCreate:
    properties:
      cart:
        properties:
          cart_id:
            type: integer
          game_id:
            type: integer
          quantity:
            type: integer
          sub_total:
            type: number
        type: object
      message:
        type: string
    type: object
  contract.ResCartDelete:
    properties:
      cart:
        properties:
          cart_id:
            type: integer
          game_id:
            type: integer
          quantity:
            type: integer
          sub_total:
            type: number
        type: object
      message:
        type: string
    type: object
  contract.ResCartEstimate:
    properties:
      created_at:
        type: string
      order_details:
        items:
          properties:
            game:
              properties:
                description:
                  type: string
                game_id:
                  type: integer
                genres:
                  items:
                    type: string
                  type: array
                price:
                  type: number
                title:
                  type: string
                year:
                  type: integer
              type: object
            quantity:
              type: integer
            sub_total:
              type: number
          type: object
        type: array
      price_total:
        type: number
    type: object
  contract.ResCartUpdate:
    properties:
      cart:
        properties:
          cart_id:
            type: integer
          game_id:
            type: integer
          quantity:
            type: integer
          sub_total:
            type: number
        type: object
      message:
        type: string
    type: object
  contract.ResGameGetAll:
    properties:
      games:
        items:
          properties:
            description:
              type: string
            game_id:
              type: integer
            genres:
              items:
                type: string
              type: array
            price:
              type: number
            stock:
              type: integer
            title:
              type: string
            year:
              type: integer
          type: object
        type: array
      metadata:
        $ref: '#/definitions/contract.Metadata'
    type: object
...
host: localhost:8080
info:
  contact:
    email: ucokkocu411@gmail.com
    name: ucok-man
  description: Documentation for PixelRental API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PixelRental API
  version: "1.0"
paths:
  /carts:
    get:
      consumes:
      - application/json
      description: Get all available carts record
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contract.ResGameGetAll'
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "429":
          description: Too Many Requests
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
      summary: Get all carts
      tags:
      - carts
    post:
      consumes:
      - application/json
      description: Create new carts record
      parameters:
      - description: Create Cart
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/contract.ReqCartCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/contract.ResCartCreate'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "429":
          description: Too Many Requests
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
      summary: Create carts
      tags:
      - carts
  /carts/:id:
    delete:
      consumes:
      - application/json
      description: Delete carts record
      parameters:
      - description: cart id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contract.ResCartDelete'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "429":
          description: Too Many Requests
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
      summary: Delete carts
      tags:
      - carts
    put:
      consumes:
      - application/json
      description: Update carts record
      parameters:
      - description: cart id
        in: path
        name: id
        required: true
        type: integer
      - description: Update Cart
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/contract.ReqCartUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contract.ResCartUpdate'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "429":
          description: Too Many Requests
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
      summary: Update carts
      tags:
      - carts
  /carts/estimate:
    get:
      consumes:
      - application/json
      description: Get estimation calculated carts record
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contract.ResCartEstimate'
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "429":
          description: Too Many Requests
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
      summary: Get calculated carts
      tags:
      - carts
...
swagger: "2.0"